Fix bug with template in combination with pandoc
If the FHS template was used in combination with
RMarkdown/pandoc then the FHS logos on the title page disappeared.

The reason was that the pandoc beamer template inserted some graphicx
width/height defaults as soon as the presentation contains graphics.

This bug fix overrides the (potentially) existing defaults for
the title page graphics.